Put winners first
All submission were an anchor tag wrapping a div. The div had the submission class on it. The winning div had a winner class on it.

Unfortunately, as the divs were wrapped in anchor tags, they were not direct children of the challenge-container div which had display: grid.

Therefore they were not 'grid' elements.

I swapped it so that the divs now wrap the anchor tags and therefore are now children of the challenge container.

It now works as expected.