Add submission count to previous challenges
I've modified the `displaySubmissionCount()` function to take two
arguments: `challengeName` & `submissionText`. This now allows you to
use this function to get the submission count for the current challenge
and any of the previous challenges.

For each previous challenge:
* The html file now contains a `<p>` element in the header for rendering
the count
* the index.js file now also calls displaySubmissionCount() to get the
count and render it to the page.

So far, I have not made changes to the Freestyle, Pricing Component, and
Tip Calculator challenges, because these have been hard coded and do not
use JS.